+++ server/kustaurant/src/main/java/com/kustaurant/kustaurant/admin/naverPlaceCrawl/exception/InvalidNaverPlaceUrlException.java
@@ -1,0 +1,8 @@
+package com.kustaurant.kustaurant.admin.naverPlaceCrawl.exception;
+
+public class InvalidNaverPlaceUrlException extends RuntimeException {
+
+    public InvalidNaverPlaceUrlException(String message) {
+        super(message);
+    }
+}
