+++ server/kustaurant/src/main/java/com/kustaurant/kustaurant/admin/naverPlaceCrawl/infrastructure/RestaurantCrawlRawRepository.java
@@ -5,5 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RestaurantCrawlRawRepository extends JpaRepository<RestaurantCrawlRawEntity, Long> {
 
+    Optional<RestaurantCrawlRawEntity> findBySourcePlaceId(String sourcePlaceId);
+
     Optional<RestaurantCrawlRawEntity> findBySourceUrl(String sourceUrl);
 }
